refactor: update MapView component to remove native implementation and streamline imports
